python/proxmox-oidc-user-sync/requirements.txt
proxmoxer
python-ldap